requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.24.3
python-jose==3.3.0
passlib==1.7.4
pydantic==2.5.0